@@ -1,6 +1,6 @@
 {
-  "name": "default",
-  "version": "1.0.0",
+  "name": "@kkt/template-markdown",
+  "version": "5.5.1",
   "description": "",
   "private": true,
   "scripts": {
@@ -14,12 +14,26 @@
   "author": "",
   "license": "MIT",
   "dependencies": {
-    "react": "^16.4.0",
-    "react-dom": "^16.4.0"
+    "react": "^16.12.0",
+    "react-dom": "^16.12.0"
   },
   "devDependencies": {
-    "@kkt/plugin-less": "^1.0.2",
-    "kkt": "file:../../",
+    "kkt": "5.5.1",
     "raw-loader": "^1.0.0"
+  },
+  "eslintConfig": {
+    "extends": "react-app"
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   }
 }
