@@ -1,7 +1,7 @@
 {
-  "name": "default",
-  "version": "1.0.0",
-  "description": "",
+  "name": "@kkt/template-rematch",
+  "version": "5.5.1",
+  "description": "Use Rematch for the project.",
   "private": true,
   "scripts": {
     "start": "kkt start",
@@ -14,18 +14,34 @@
   "author": "",
   "license": "MIT",
   "dependencies": {
-    "@rematch/core": "^1.0.6",
-    "axios": "^0.18.0",
-    "history": "^4.7.2",
-    "react": "^16.7.0",
-    "react-dom": "^16.7.0",
-    "react-dynamic-loadable": "^1.1.2",
-    "react-redux": "^6.0.0",
-    "react-router-dom": "^4.3.1"
+    "@rematch/core": "^1.3.0",
+    "@rematch/loading": "^1.2.1",
+    "axios": "^0.19.0",
+    "history": "^4.10.1",
+    "react": "^16.12.0",
+    "react-dom": "^16.12.0",
+    "react-dynamic-loadable": "^2.0.1",
+    "react-redux": "^7.1.3",
+    "react-router-dom": "^5.1.2",
+    "redux": "4.0.4"
   },
   "devDependencies": {
-    "@kkt/plugin-less": "^1.0.2",
-    "kkt": "file:../../",
-    "mocker-api": "^1.6.6"
+    "@kkt/loader-less": "5.5.1",
+    "kkt": "5.5.1"
+  },
+  "eslintConfig": {
+    "extends": "react-app"
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   }
 }
