@@ -1,7 +1,7 @@
 {
-  "name": "default",
-  "version": "1.0.0",
-  "description": "",
+  "name": "@kkt/template-scss",
+  "version": "5.5.1",
+  "description": "Use an example of SCSS.",
   "private": true,
   "scripts": {
     "start": "kkt start",
@@ -14,11 +14,26 @@
   "author": "",
   "license": "MIT",
   "dependencies": {
-    "react": "^16.4.0",
-    "react-dom": "^16.4.0"
+    "react": "^16.12.0",
+    "react-dom": "^16.12.0"
   },
   "devDependencies": {
-    "kkt": "file:../../",
-    "@kkt/plugin-scss": "^1.0.3"
+    "@kkt/loader-scss": "5.5.1",
+    "kkt": "5.5.1"
+  },
+  "eslintConfig": {
+    "extends": "react-app"
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   }
 }
