@@ -1,14 +1,11 @@
 {
-  "name": "typescript-example",
-  "version": "1.0.0",
-  "description": "typescript example",
+  "name": "@kkt/template-typescript",
+  "version": "5.5.1",
+  "description": "Use an example of TypeScript.",
   "private": true,
   "scripts": {
     "start": "kkt start",
     "build": "kkt build",
-    "bundle": "kkt build --bundle --no-emptyDir",
-    "bundle:watch": "kkt build --bundle --no-emptyDir -w",
-    "bundle:min": "kkt build --bundle=min --no-emptyDir",
     "test": "kkt test --env=jsdom",
     "test:coverage": "kkt test --env=jsdom --coverage"
   },
@@ -20,15 +17,30 @@
   "license": "MIT",
   "dependencies": {
     "classnames": "^2.2.6",
-    "react": "^16.4.0",
-    "react-dom": "^16.4.0"
+    "react": "^16.12.0",
+    "react-dom": "^16.12.0"
   },
   "devDependencies": {
-    "@kkt/plugin-less": "^1.2.0",
+    "@kkt/loader-less": "5.5.1",
     "@types/classnames": "^2.2.8",
-    "@types/jest": "^24.0.13",
-    "@types/react": "^16.8.19",
-    "@types/react-dom": "^16.8.4",
-    "kkt": "file:../../"
+    "@types/jest": "^24.0.23",
+    "@types/react": "^16.9.13",
+    "@types/react-dom": "^16.9.4",
+    "kkt": "5.5.1"
+  },
+  "eslintConfig": {
+    "extends": "react-app"
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   }
 }
