@@ -19,14 +19,12 @@ const App = () => (
       >
         Learn React
       </a>
-      <div>
-        <Button icon="weibo" type="primary">主要按钮</Button>
-        <Button icon="chrome" type="success">成功按钮</Button>
-        <Button icon="taobao" type="warning">警告按钮</Button>
-        <Button icon="weibo" type="danger" />
+      <div className={styles.btn}>
+        <Button type="primary">主要按钮</Button>
+        <Button type="success">成功按钮</Button>
+        <Button type="warning">警告按钮</Button>
         <Button type="danger">错误按钮</Button>
-        <Button icon="chrome" type="light">亮按钮</Button>
-        <Button icon="apple" type="dark">暗按钮</Button>
+        <Button type="light">亮按钮</Button>
         <Button type="dark">暗按钮</Button>
       </div>
     </header>
