@@ -1,3 +1,3 @@
-import createHistory from 'history/createBrowserHistory';
+// import createBrowserHistory from 'history/createBrowserHistory';
 
-export default createHistory();
+export default require('history').createBrowserHistory();
